+++ src/Auth.spec.ts
@@ -10,8 +10,6 @@ import {
 	isDefined,
 	hasConstructor,
 	hasMethod,
-	hasProperty,
-	extendsClass,
 	hasDefaultExport,
 	reexports,
 	hasEnumeral,
@@ -19,13 +17,14 @@ import {
 } from "./test/JasmineExtender";
 import * as Utils from "./Utils";
 import AbstractContext from "./AbstractContext";
-import * as App from "./App";
 import AuthenticationToken from "./Auth/AuthenticationToken";
 import Authenticator from "./Auth/Authenticator";
 import BasicAuthenticator from "./Auth/BasicAuthenticator";
 import * as Token from "./Auth/Token";
 import TokenAuthenticator from "./Auth/TokenAuthenticator";
 import UsernameAndPasswordToken from "./Auth/UsernameAndPasswordToken";
+import UsernameAndPasswordCredentials from "./Auth/UsernameAndPasswordCredentials";
+import * as Errors from "./Errors";
 
 import * as Auth from "./Auth";
 import DefaultExport from "./Auth";
@@ -124,7 +123,18 @@ describe( module( "Carbon/Auth" ), ():void => {
 		});
 
 		it( hasConstructor(), ():void => {
-			// TODO test
+			let context:AbstractContext;
+			class MockedContext extends AbstractContext {
+				resolve( uri:string ) {
+					return uri;
+				}
+			}
+			context = new MockedContext();
+
+			let auth = new Auth.Class( context );
+
+			expect( auth ).toBeTruthy();
+			expect( auth instanceof Auth.Class ).toBe( true );
 		});
 
 		it( hasMethod(
@@ -138,28 +148,286 @@ describe( module( "Carbon/Auth" ), ():void => {
 			// TODO test
 		});
 
+		it( hasMethod(
+			INSTANCE,
+			"authenticate",
+			"Authenticate the user with an `username` and `password`. Uses the `TOKEN` method for the authentication.", [
+				{ name: "username", type: "string" },
+				{ name: "password", type: "string" }
+			],
+			{ type: "Promise<Carbon.Auth.Credentials>" }
+		), ():void => {
+			class MockedContext extends AbstractContext {
+				resolve( uri:string ) {
+					return uri;
+				}
+			}
+			let context = new MockedContext();
+
+			let auth = new Auth.Class( context );
+
+			expect( auth.authenticate ).toBeDefined();
+			expect( Utils.isFunction( auth.authenticate ) ).toBe( true );
+
+			let spy = spyOn( auth, "authenticateUsing" );
+
+			auth.authenticate( "myUer@user.com", "myAwesomePassword" );
+
+			expect( spy ).toHaveBeenCalledWith( "TOKEN", "myUer@user.com", "myAwesomePassword" );
+		});
+
 		describe( method(
 			INSTANCE,
-			"authenticate"
+			"authenticateUsing"
 		), ():void => {
+			let context:AbstractContext;
 
-			it( hasSignature(
-				"Authenticate the user with an `username` and `password`.", [
-					{ name: "username", type: "string" },
-					{ name: "password", type: "string" }
-				],
-				{ type: "Promise<void>" }
-			), ():void => {
-				// TODO test
+			beforeEach( ():void => {
+				class MockedContext extends AbstractContext {
+					resolve( uri:string ) {
+						return uri;
+					}
+				}
+				context = new MockedContext();
+
+				jasmine.Ajax.install();
+			});
+
+			afterEach( ():void => {
+				jasmine.Ajax.uninstall();
 			});
 
 			it( hasSignature(
-				"Authenticate the user with an instance of `Carbon.Auth.AuthenticationToken`.", [
-					{ name: "authenticationToken", type: "Carbon.Auth.AuthenticationToken" },
+				"Authenticates the user with Basic HTTP Authentication, witch uses encoded username and password.", [
+					{ name: "method", type: "'BASIC'" },
+					{ name: "username", type: "string" },
+					{ name: "password", type: "string" }
 				],
-				{ type: "Promise<void>" }
-			), ():void => {
-				// TODO test
+				{ type: "Promise<Carbon.Auth.UsernameAndPasswordCredentials.Class>"}
+			), ( done:{ ():void, fail:() => void } ):void => {
+				let auth01 = new Auth.Class( context );
+				
+				expect( auth01.authenticateUsing ).toBeDefined();
+				expect( Utils.isFunction( auth01.authenticateUsing ) ).toBe( true );
+				expect( auth01.isAuthenticated() ).toBe( false );
+
+				let username = "myUser@user.com";
+				let password = "myAwesomePassword";
+				let promises:Promise<any>[] = [];
+				let promise:Promise<any>;
+				let spies = {
+					success: ( credentials:UsernameAndPasswordCredentials ):void => {
+						expect( auth01.isAuthenticated() ).toBe( true );
+						expect( credentials.username ).toEqual( username );
+						expect( credentials.password ).toEqual( password );
+					},
+					fail: ( error ):void => {
+						expect( error.name ).toBe( Errors.IllegalArgumentError.name );
+					}
+				};
+				let spySuccess = spyOn( spies, "success" ).and.callThrough();
+				let spyFail = spyOn( spies, "fail" ).and.callThrough();
+
+				promise = auth01.authenticateUsing( "BASIC", username, password );
+				expect( promise instanceof Promise ).toBe( true );
+				promises.push( promise.then( spies.success, spies.fail ) );
+
+
+				let auth02 = new Auth.Class( context );
+				promise = auth02.authenticateUsing( "BASIC", {} );
+				expect( promise instanceof Promise ).toBe( true );
+				promises.push( promise.then( spies.success, spies.fail ) );
+
+				let auth03 = new Auth.Class( context );
+				promise = auth03.authenticateUsing( "Error", username, password );
+				expect( promise instanceof Promise ).toBe( true );
+				promises.push( promise.then( spies.success, spies.fail ) );
+
+				Promise.all( promises ).then( ():void => {
+					expect( spySuccess.calls.count() ).toBe( 1 );
+					expect( spyFail.calls.count() ).toBe( 2 );
+					done();
+				}, done.fail );
+			});
+
+			it( hasSignature(
+				"Authenticates the user with username and password, and generates a JSON Web Token (JWT) credentials.", [
+					{ name: "method", type: "'TOKEN'" },
+					{ name: "username", type: "string" },
+					{ name: "password", type: "string" }
+				],
+				{ type: "Promise<Carbon.Auth.Token.Class>"}
+			), ( done:{ ():void, fail:() => void } ):void => {
+				let auth01 = new Auth.Class( context );
+
+				expect( auth01.authenticateUsing ).toBeDefined();
+				expect( Utils.isFunction( auth01.authenticateUsing ) ).toBe( true );
+				expect( auth01.isAuthenticated() ).toBe( false );
+
+				let date:Date = new Date();
+				date.setDate( date.getDate() + 1 );
+				let token:Object = [
+					{
+						"@id": "_:BlankNode",
+						"@type": [
+							"https://carbonldp.com/ns/v1/security#Token",
+							"https://carbonldp.com/ns/v1/platform#VolatileResource"
+						],
+						"https://carbonldp.com/ns/v1/security#expirationTime": [
+							{
+								"@type": "http://www.w3.org/2001/XMLSchema#dateTime",
+								"@value": date.toISOString()
+							}
+						],
+						"https://carbonldp.com/ns/v1/security#tokenKey": [
+							{
+								"@value": "eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJodHRwczovL2V4YW1wbGUuY29tL3VzZXJzL2EtdXNlci8iLCJleHAiOjExNDkxMjAwMDAwMDB9.T8XSFKyiT-5PAx_yxv2uY94nfvx65zWz8mI2OlSUFnE"
+							}
+						]
+					}
+				];
+				let promises:Promise<any>[] = [];
+				let promise:Promise<any>;
+				let spies = {
+					success: ( credentials:Token.Class ):void => {
+						expect( auth01.isAuthenticated() ).toBe( true );
+						expect( credentials.key ).toEqual( token[0]["https://carbonldp.com/ns/v1/security#tokenKey"][0]["@value"] );
+					},
+					fail: ( error ):void => {
+						expect( error.name ).toBe( Errors.IllegalArgumentError.name );
+					}
+				};
+				let spySuccess = spyOn( spies, "success" ).and.callThrough();
+				let spyFail = spyOn( spies, "fail" ).and.callThrough();
+
+				promise = auth01.authenticateUsing( "TOKEN", "myUser@user.con", "myAwesomePassword" );
+				expect( promise instanceof Promise ).toBe( true );
+				promises.push( promise.then( spies.success, spies.fail ) );
+
+				jasmine.Ajax.stubRequest( /token/ ).andReturn({
+					status: 200,
+					responseText: JSON.stringify( token )
+				});
+
+				let auth02:Auth.Class = new Auth.Class( context );
+				promise = auth02.authenticateUsing( "TOKEN", {} );
+				expect( promise instanceof Promise ).toBe( true );
+				promises.push( promise.then( spies.success, spies.fail ) );
+
+				let auth03:Auth.Class = new Auth.Class( context );
+				promise = auth03.authenticateUsing( "Error", "myUser@user.con", "myAwesomePassword" );
+				expect( promise instanceof Promise ).toBe( true );
+				promises.push( promise.then( spies.success, spies.fail ) );
+
+				Promise.all( promises ).then( ():void => {
+					expect( spySuccess.calls.count() ).toBe( 1 );
+					expect( spyFail.calls.count() ).toBe( 2 );
+					done();
+				}, done.fail );
+			});
+
+			it( hasSignature(
+				"Authenticates the user with a JSON Web Token (JWT), i.e. the credentials generated by TokenAuthenticator.", [
+					{ name: "method", type: "'TOKEN'" },
+					{ name: "token", type: "Carbon.Auth.Token.Class" }
+				],
+				{ type: "Promise<Carbon.Auth.Token.Class>" }
+			), ( done:{ ():void, fail:() => void } ):void => {
+				let auth01 = new Auth.Class( context );
+				let auth02 = new Auth.Class( context );
+				
+				expect( auth01.authenticateUsing ).toBeDefined();
+				expect( Utils.isFunction( auth01.authenticateUsing ) ).toBe( true );
+				expect( auth01.isAuthenticated() ).toBe( false );
+
+
+				let date:Date;
+				let token:Token.Class;
+				let promises:Promise<any>[] = [];
+				let promise:Promise<any>;
+				let spies = {
+					success01: ( credential:Token.Class ):void => {
+						expect( credential.key ).toEqual( token.key );
+						expect( auth01.isAuthenticated() ).toBe( true );
+					},
+					success02: ( credential:Token.Class ):void => {
+						expect( credential.key ).toEqual( token.key );
+						expect( auth02.isAuthenticated() ).toBe( true );
+					},
+					fail: ( error ):void => {
+						expect( error.name ).toBe( Errors.IllegalArgumentError.name );
+					}
+				};
+				let spySuccess01 = spyOn( spies, "success01" ).and.callThrough();
+				let spySuccess02 = spyOn( spies, "success02" ).and.callThrough();
+				let spyFail = spyOn( spies, "fail" ).and.callThrough();
+
+				// OK Token
+				date = new Date();
+				date.setDate( date.getDate() + 1 );
+				token = <any> {
+					expirationTime: date,
+					id: "_:BlankNode",
+					key: "eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJodHRwczovL2V4YW1wbGUuY29tL3VzZXJzL2EtdXNlci8iLCJleHAiOjExNDkxMjAwMDAwMDB9.T8XSFKyiT-5PAx_yxv2uY94nfvx65zWz8mI2OlSUFnE",
+					types: [
+						"https://carbonldp.com/ns/v1/security#Token",
+						"https://carbonldp.com/ns/v1/platform#VolatileResource"
+					]
+				};
+				promise = auth01.authenticateUsing( "TOKEN", token );
+				expect( promise instanceof Promise ).toBe( true );
+				promises.push( promise.then( spies.success01, spies.fail ) );
+
+				// Will be OK, if expirationDate is a string the method tries parse it to a Date object
+				date = new Date();
+				date.setDate( date.getDate() + 1 );
+				var getFromStorage = ():Object => {
+					let storedToken = {
+						expirationTime: date.toISOString(),
+						id: "_:BlankNode",
+						key: "eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJodHRwczovL2V4YW1wbGUuY29tL3VzZXJzL2EtdXNlci8iLCJleHAiOjExNDkxMjAwMDAwMDB9.T8XSFKyiT-5PAx_yxv2uY94nfvx65zWz8mI2OlSUFnE",
+						types: [
+							"https://carbonldp.com/ns/v1/security#Token",
+							"https://carbonldp.com/ns/v1/platform#VolatileResource"
+						]
+					};
+					return JSON.parse( JSON.stringify( storedToken ) );
+				};
+				promise = auth02.authenticateUsing( "TOKEN", <Token.Class> getFromStorage() );
+				expect( promise instanceof Promise ).toBe( true );
+				promises.push( promise.then( spies.success02, spies.fail ) );
+
+				// Will fail, because expirationDate has been reached
+				let auth03:Auth.Class = new Auth.Class( context );
+				date = new Date();
+				date.setDate( date.getDate() - 1 );
+				token = <any> {
+					expirationTime: date,
+					id: "_:BlankNode",
+					key: "eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJodHRwczovL2V4YW1wbGUuY29tL3VzZXJzL2EtdXNlci8iLCJleHAiOjExNDkxMjAwMDAwMDB9.T8XSFKyiT-5PAx_yxv2uY94nfvx65zWz8mI2OlSUFnE",
+					types: [
+						"https://carbonldp.com/ns/v1/security#Token",
+						"https://carbonldp.com/ns/v1/platform#VolatileResource"
+					]
+				};
+				promise = auth03.authenticateUsing( "TOKEN", token );
+				expect( promise instanceof Promise ).toBe( true );
+				promises.push( promise.then( spies.success01, spies.fail ) );
+
+				promise = auth03.authenticateUsing( "TOKEN", {} );
+				expect( promise instanceof Promise ).toBe( true );
+				promises.push( promise.then( spies.success01, spies.fail ) );
+
+				promise = auth03.authenticateUsing( "Error", token );
+				expect( promise instanceof Promise ).toBe( true );
+				promises.push( promise.then( spies.success01, spies.fail ) );
+
+				Promise.all( promises ).then( ():void => {
+					expect( spySuccess01.calls.count() ).toBe( 1 );
+					expect( spySuccess02.calls.count() ).toBe( 1 );
+					expect( spyFail.calls.count() ).toBe( 3 );
+					done();
+				}, done.fail );
 			});
 
 		});